- firebird support.  reflection works fully, overall test success in the 75% range approx - oracle and firebird now normalize column names to SQLA "lowercase" for result.keys() allowing consistent identifier name visibility on the client side.

+++ lib/sqlalchemy/types.py
@@ -923,7 +923,7 @@ class SMALLINT(SmallInteger):
     __visit_name__ = 'SMALLINT'
 
 
-class BIGINT(SmallInteger):
+class BIGINT(BigInteger):
     """The SQL BIGINT type."""
 
     __visit_name__ = 'BIGINT'
